perf(properties): auth & validate before delete

+++ backend-fastify/src/controllers/properties/delete-property.js
@@ -1,10 +1,14 @@
 import { Property } from "../../models/property.js";
+import { authBearerToken } from "../../utils/requests.js";
+import { userIdToken } from "../../utils/users.js";
 import { unlinkImages } from "./image-property.js";
 
 export const deleteProperty = async function (req, res) {
   const { id } = req.params;
   try {
-    const property = await Property.findOneAndDelete({ property_id: id });
+    const token = authBearerToken(req);
+    const user_id = userIdToken(token);
+    const property = await Property.findOneAndDelete({ property_id: id, user_id });
     if (!property) {
       res.status(404).send({});
     }
